HTMLParser: Link for attributes with unique input id and fixing memory leaks (Part2-F)
used a list that holds a pointer to adress of a node with ids of an input so we can only compare ids of inputes with our fors and save time,
as we only care about linking a for with the id of an input.Fixed memory leaks using valgrind.
also fixed a double free bug with type and value what happened was '' and "" return both quoted string as a tokken
now because in flex in the '' quote rule we dont do strdup the freed value from "" (in the type field) remained in yylval.str
meaning we try to free a null pointer.i fixed the issue by just doing strdup in the single quotes rule.

[LinkBoi00] adress->address, drop old memleak fixes, added free($4) to input_attributes because linked_list now has strdup

Signed-off-by: Elias Dimopoulos <linkdevel@protonmail.com>