HTMLParser:adding relative and absolute urls
href can now take an absolute and a relative url.If
it has a # it checks to see if its linked with an id before it in the page.
src now also take a relative or an absolute url.
if we see // //: or a white space then the url is invalid