HTMLParser: Adding relative and absolute urls (Part2-C,D)
href can now take an absolute and a relative url.If
it has a # it checks to see if its linked with an id before it in the page.
src now also take a relative or an absolute url.
if we see // //: or a white space then the url is invalid

[LinkBoi00] Move https/http logic to validate_url

Signed-off-by: Elias Dimopoulos <linkdevel@protonmail.com>